chore: move collector, middleware and enum to test level dir
